requirements-light.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
librosa==0.9.2
tensorflow-cpu==2.12.0
scikit-learn==1.2.2
numpy==1.23.5
pydub==0.25.1
soundfile==0.12.1
